fix: improve html escaping of element attributes
