+++ .changeset/young-ads-roll.md
@@ -1,0 +1,5 @@
+---
+"svelte": patch
+---
+
+fix: improve html escaping of element attributes
