+++ packages/svelte/tests/runtime-legacy/samples/escaped-attr-2/main.svelte
@@ -1,0 +1,8 @@
+<script>
+	const x = `</noscript><script>console.log('should not run')<` + `/script>`
+</script>
+
+<noscript>
+	<a href={x}>test</a>
+</noscript>
+
