+++ packages/svelte/tests/runtime-legacy/samples/escaped-attr/_config.js
@@ -1,0 +1,7 @@
+import { test } from '../../test';
+
+export default test({
+	test({ assert, logs }) {
+		assert.deepEqual(logs, []);
+	}
+});
