+++ packages/svelte/tests/runtime-legacy/samples/escaped-attr/main.svelte
@@ -1,0 +1,4 @@
+<noscript>
+	<a href="</noscript><script>console.log('should not run')</script>">test</a>
+</noscript>
+
